ai-engine-cpu/requirements.txt
matplotlib==3.7.1
numpy==1.24.2
onnxruntime
paho_mqtt==2.1.0
Pillow==10.4.0
PyYAML==6.0.2
tqdm==4.65.0
opencv_python==4.7.0.72